APT-235: UDP-connection related functionality moved to core.h to be used for all supported packet types
