+++ .gitignore
@@ -2,3 +2,4 @@ build/
 .vscode/
 Testing/
 *.dump
+*.ctu-info
