+++ CMakeLists.txt
@@ -214,7 +214,7 @@ else()
 endif()
 
 # Define the core library
-add_library(provizio_radar_api_core STATIC src/common.c src/socket.c src/radar_point_cloud.c src/util.c)
+add_library(provizio_radar_api_core STATIC src/common.c src/socket.c src/radar_point_cloud.c src/util.c src/core.c)
 target_include_directories(provizio_radar_api_core PUBLIC include)
 set_property(TARGET provizio_radar_api_core PROPERTY C_STANDARD 99) # As defined by MISRA
 if(WIN32)
