+++ include/provizio/radar_api/radar_point_cloud.h
@@ -173,11 +173,6 @@ typedef struct provizio_radar_point_cloud_api_context
     provizio_radar_point_cloud_api_context_impl impl;
 } provizio_radar_point_cloud_api_context;
 
-typedef struct provizio_radar_point_cloud_api_connection
-{
-    PROVIZIO__SOCKET sock;
-} provizio_radar_point_cloud_api_connection;
-
 /**
  * @brief Initializes a provizio_radar_point_cloud_api_context object to handle a single radar
  *
@@ -221,8 +216,8 @@ PROVIZIO__EXTERN_C int32_t provizio_radar_point_cloud_api_context_assign(
  * @param context Previously initialized provizio_radar_point_cloud_api_context
  * @param packet Valid provizio_radar_point_cloud_packet
  * @param packet_size The size of the packet, to check data is valid and avoid out-of-bounds access
- * @return 0 in case the packet was handled successfully, EAGAIN in case the packet was skipped as obsolete, other error
- * code in case of another error
+ * @return 0 in case the packet was handled successfully, PROVIZIO_E_SKIPPED in case the packet was skipped as obsolete,
+ * other error code in case of another error
  *
  * @warning radar_position_id of all packets handled by this context must be same (returns an error otherwise)
  */
@@ -237,8 +232,8 @@ PROVIZIO__EXTERN_C int32_t provizio_handle_radar_point_cloud_packet(provizio_rad
  * @param num_contexts Number of contexts (i.e. max numbers of radars to handle)
  * @param packet Valid provizio_radar_point_cloud_packet
  * @param packet_size The size of the packet, to check data is valid and avoid out-of-bounds access
- * @return 0 in case the packet was handled successfully, EAGAIN in case the packet was skipped as obsolete, EBUSY in
- * case num_contexts is not enough, other error code in case of another error
+ * @return 0 in case the packet was handled successfully, PROVIZIO_E_SKIPPED in case the packet was skipped as obsolete,
+ * PROVIZIO_E_OUT_OF_CONTEXTS in case num_contexts is not enough, other error code in case of another error
  */
 PROVIZIO__EXTERN_C int32_t provizio_handle_radars_point_cloud_packet(provizio_radar_point_cloud_api_context *contexts,
                                                                      size_t num_contexts,
@@ -252,8 +247,8 @@ PROVIZIO__EXTERN_C int32_t provizio_handle_radars_point_cloud_packet(provizio_ra
  * @param context Previously initialized provizio_radar_point_cloud_api_context
  * @param payload The payload of the UDP packet
  * @param payload_size The size of the payload in bytes
- * @return 0 if it's a provizio_radar_point_cloud_packet and it was handled successfully, EAGAIN if it's not a
- * provizio_radar_point_cloud_packet, other error code if it's a provizio_radar_point_cloud_packet but its handling
+ * @return 0 if it's a provizio_radar_point_cloud_packet and it was handled successfully, PROVIZIO_E_SKIPPED if it's not
+ * a provizio_radar_point_cloud_packet, other error code if it's a provizio_radar_point_cloud_packet but its handling
  * failed for another reason
  *
  * @warning if it's a provizio_radar_point_cloud_packet, radar_position_id of all packets handled by this context must
@@ -270,73 +265,15 @@ PROVIZIO__EXTERN_C int32_t provizio_handle_possible_radar_point_cloud_packet(
  * @param num_contexts Number of contexts (i.e. max numbers of radars to handle)
  * @param payload The payload of the UDP packet
  * @param payload_size The size of the payload in bytes
- * @return 0 if it's a provizio_radar_point_cloud_packet and it was handled successfully, EAGAIN if it's not a
- * provizio_radar_point_cloud_packet, EBUSY in case num_contexts is not enough, other error code if it's a
- * provizio_radar_point_cloud_packet but its handling failed for another reason
+ * @return 0 if it's a provizio_radar_point_cloud_packet and it was handled successfully, PROVIZIO_E_SKIPPED if it's not
+ * a provizio_radar_point_cloud_packet, PROVIZIO_E_OUT_OF_CONTEXTS in case num_contexts is not enough, other error code
+ * if it's a provizio_radar_point_cloud_packet but its handling failed for another reason
  *
  * @warning if it's a provizio_radar_point_cloud_packet, radar_position_id of all packets handled by this context must
  * be same (returns an error otherwise)
  */
 PROVIZIO__EXTERN_C int32_t provizio_handle_possible_radars_point_cloud_packet(
     provizio_radar_point_cloud_api_context *contexts, size_t num_contexts, const void *payload, size_t payload_size);
-
-#define PROVIZIO__RADAR_API_POINT_CLOUD_DEFAULT_PORT ((uint16_t)7769)
-
-/**
- * @brief Connect to the radar point clouds API to start receiving packets by UDP
- *
- * @param udp_port UDP port to receive from, by default = PROVIZIO__RADAR_API_POINT_CLOUD_DEFAULT_PORT
- * @param receive_timeout_ns Max number of nanoseconds provizio_radar_point_cloud_api_receive_packet should wait for a
- * packet, or 0 to wait as long as required
- * @param check_connection Use any non-zero value if the connection is to be checked to be receiving anything prior to
- * returning a successful result
- * @param out_connection A provizio_radar_point_cloud_api_connection to store the connection handle
- * @return 0 if received successfully, EAGAIN if timed out, other error value if failed for another reason
- */
-PROVIZIO__EXTERN_C int32_t
-provizio_radar_point_cloud_api_open_connection(uint16_t udp_port, uint64_t receive_timeout_ns, uint8_t check_connection,
-                                               provizio_radar_point_cloud_api_connection *out_connection);
-
-/**
- * @brief Receive and handle the next UDP packet from a single radar using a previously connected API and a previously
- * initialized provizio_radar_point_cloud_api_context
- *
- * @param context provizio_radar_point_cloud_api_context previously initialized with
- * provizio_radar_point_cloud_api_context_init
- * @param connection A provizio_radar_point_cloud_api_connection previously connected with
- * provizio_radar_point_cloud_api_open_connection
- * @return 0 if received successfully, EAGAIN if timed out, other error value if failed for another reason
- *
- * @warning radar_position_id of all packets handled by this context must be same (returns an error otherwise)
- */
-PROVIZIO__EXTERN_C int32_t provizio_radar_point_cloud_api_context_receive_packet(
-    provizio_radar_point_cloud_api_context *context, provizio_radar_point_cloud_api_connection *connection);
-
-/**
- * @brief Receive and handle the next UDP packet from one of multiple radars using a previously connected API and a
- * previously initialized array of provizio_radar_point_cloud_api_context objects
- *
- * @param contexts Array of provizio_radar_point_cloud_api_context objects previously initialized with
- * provizio_radar_point_cloud_api_contexts_init
- * @param num_contexts Number of objects in contexts
- * @param connection A provizio_radar_point_cloud_api_connection previously connected with
- * provizio_radar_point_cloud_api_open_connection
- * @return 0 if received successfully, EAGAIN if timed out or the received packet was obsolete, EBUSY in case
- * num_contexts is not enough, other error value if failed for another reason
- */
-PROVIZIO__EXTERN_C int32_t provizio_radar_point_cloud_api_contexts_receive_packet(
-    provizio_radar_point_cloud_api_context *contexts, size_t num_contexts,
-    provizio_radar_point_cloud_api_connection *connection);
-
-/**
- * @brief Closes a previously connected radar point clouds API
- *
- * @param connection A provizio_radar_point_cloud_api_connection previously connected with
- * provizio_radar_point_cloud_api_open_connection
- * @return 0 if successfull, error code otherwise
- */
-PROVIZIO__EXTERN_C int32_t
-provizio_radar_point_cloud_api_close_connection(provizio_radar_point_cloud_api_connection *connection);
 
 #if defined(__cplusplus) && __cplusplus >= 201103L
 static_assert(offsetof(provizio_radar_point_cloud_packet_protocol_header, packet_type) == 0,
