+++ src/util.c
@@ -45,7 +45,7 @@
         }                                                                                                              \
     } while (0)
 
-uint64_t provizio_ntohll(uint64_t v)
+uint64_t provizio_ntohll(uint64_t value)
 {
     const size_t bits_32 = 32;
     const uint64_t bits_32_mask = 0xFFFFFFFF;
@@ -57,17 +57,17 @@ uint64_t provizio_ntohll(uint64_t v)
         // lcov excluded as it's host CPU arch dependent
         // linting disabled as ntohl implementation is up to a platform (it uses asm instructions in some platforms,
         // which clang-tidy hates)
-        return ((uint64_t)ntohl((uint32_t)(v & bits_32_mask)) << bits_32) | // NOLINT // LCOV_EXCL_LINE
-               ntohl((uint32_t)(v >> bits_32));                             // NOLINT // LCOV_EXCL_LINE
+        return ((uint64_t)ntohl((uint32_t)(value & bits_32_mask)) << bits_32) | // NOLINT // LCOV_EXCL_LINE
+               ntohl((uint32_t)(value >> bits_32));                             // NOLINT // LCOV_EXCL_LINE
     }
 
-    return v; // LCOV_EXCL_LINE: host CPU arch dependent
+    return value; // LCOV_EXCL_LINE: host CPU arch dependent
 }
 
-uint64_t provizio_htonll(uint64_t v)
+uint64_t provizio_htonll(uint64_t value)
 {
     // As reversing, when required, is a symmetrical operation
-    return provizio_ntohll(v);
+    return provizio_ntohll(value);
 }
 
 void provizio_set_protocol_field_uint8_t(uint8_t *field, uint8_t value)
