+++ test/c_99/CMakeLists.txt
@@ -45,7 +45,7 @@ else(WIN32)
   set(MATH_LIB m)
 endif(WIN32)
 
-add_executable(provizio_radar_api_core_test_c_99 src/test_main.c src/test_common.c src/test_util.c src/test_radar_point_cloud.c)
+add_executable(provizio_radar_api_core_test_c_99 src/test_main.c src/test_common.c src/test_util.c src/test_radar_point_cloud.c src/test_core.c)
 target_link_libraries(provizio_radar_api_core_test_c_99 provizio_radar_api_core unity_testing ${MATH_LIB} ${PTHREAD_LIB})
 set_property(TARGET provizio_radar_api_core_test_c_99 PROPERTY C_STANDARD 99) # As defined by MISRA
 add_dependencies(provizio_radar_api_core_test_c_99 unity)
