+++ test/c_99/src/test_main.c
@@ -27,6 +27,7 @@ void tearDown(void) // NOLINT: this exact name is required by the unit testing f
 int provizio_run_test_common(void);
 int provizio_run_test_util(void);
 int provizio_run_test_radar_point_cloud(void);
+int provizio_run_test_core(void);
 
 int main(int argc, char *argv[])
 {
@@ -37,6 +38,7 @@ int main(int argc, char *argv[])
     result = result ? result : provizio_run_test_common();
     result = result ? result : provizio_run_test_util();
     result = result ? result : provizio_run_test_radar_point_cloud();
+    result = result ? result : provizio_run_test_core();
 
     return result;
 }
